test(wrangling): already 'wrangled' trial responses get returned as is
The difference between unwrangled and wrangled responses is
- stim levels (values) get converted to indices
- columns get renamed

If you pass in already wrangled data, that means that the stim levels are indices, which are also just valid values. The wrangling code automatically extracts the stim levels. In this case, it extracts the index values, and then maps them to index values.... themselves.

For the column renaming, this does mean that you have to pass in the correct `pair_names` and `dim_names`: dimension names can be whatever anyway (e.g., `dimA` or `target_luminance`). The `pair_names` get wrangled to single letter names (e.g., `left` -> `l`). So, when passing in already wrangled data, the names have to already match this.